### 2018-08-02 #867 @bovee
Looks interesting. I'd be happy to review.
