### 2018-08-04 #869 @whedon
[REVIEW]: js-emass: A flexible JavaScript implementation of the emass algorithm
**Submitting author:** @emptyport (<a href="http://orcid.org/0000-0002-1148-6603">Michael Porter</a>)
**Repository:** <a href="https://github.com/emptyport/js-emass" target ="_blank">https://github.com/emptyport/js-emass</a>
**Version:** 1.2.4
