### 2018-08-07 emptyport/js-emass#1 @bovee
Tag and pin a 1.2.4 release on GitHub
As part of the [JOSS Paper Review](https://github.com/openjournals/joss-reviews/issues/869), it would be great to have a pinned release on GitHub to match the one on NPM (and the one submitted with the paper). This should just entail tagging the 1.2.4 (last?) commit, pushing the tag to GitHub, and creating a release from on the repo's page.
