### 2018-08-07 emptyport/js-emass#3 @bovee
Installation and Quickstart enhancements
The installation and quick start directions are good (I was able to follow along through the whole thing), but there are a couple small things I think would help to enhance them:

 - [x] The quickstart directions require `molecular-formula`; the installation directions should probably include it too, i.e. `npm install js-emass molecular-formula --save`.
 - [x] It would be useful to state that the quickstart directions are executed in a node terminal (e.g. I'm not sure the average mass spectrometrist would know how to execute these commands).
 - [x] The output from `console.log(formula.composition);` isn't in the example output.
 - [x] Another, more applied example would be great; something like distinguishing CO fragments from N2? Or something more proteomic? I think this would help establish the need for the software too.
