### 2018-08-07 emptyport/js-emass#4 @bovee
Contribution guidelines
It would be nice to have a section in the README about how contributions are handled (via pull request?), how issues are reported (on Github?) and where to find support (via Github issues? via email?)
