### 2018-08-07 #869 @arfon
> This was a pleasure to review. Overall I think this software is a clean, well-implemented mass isotope calculator with good tests, clear API documentation and an easy installation.

Thanks @bovee! I notice there are a few unchecked boxes here still, are they all items that are outstanding in your mind? If so, would you mind linking to the issues you've made on the https://github.com/emptyport/js-emass repository here?
