### 2018-08-07 #869 @bovee
Yes, and sorry, they're all linked to issues I opened:

 - Setting a defined revision in the git repository for when this was reviewed (https://github.com/emptyport/js-emass/issues/1)
 - A few things to improve the usability of the installation and quickstart directions (https://github.com/emptyport/js-emass/issues/3)
 - Adding contribution guidelines (https://github.com/emptyport/js-emass/issues/4)

And I'd also like to see more elaboration on what specific problems the software solves in the paper so the "statement of purpose" is more compelling to someone who's not familiar with the original algorithm and software (but I'm not sure if that's an issue to open on the repo or not).
