### 2018-08-07 emptyport/js-emass#3 @emptyport
All done. Will mark complete once I push the changes.
