### 2018-08-07 emptyport/js-emass#4 @emptyport
Done. Will mark closed after I push changes.
