### 2018-08-07 03e93c7
making changes after JOSS review
