+++ paper/paper.bib
@@ -6,20 +6,20 @@
     pages = "415 - 419",
     year = "2006",
     issn = "1044-0305",
-    doi = "https://doi.org/10.1016/j.jasms.2005.12.001",
+    doi = "10.1016/j.jasms.2005.12.001",
     url = "http://www.sciencedirect.com/science/article/pii/S1044030505010160",
     author = "Alan L. Rockwood and Perttu Haimi"
 }
 
 @article{DEUTERATER,
-author = {Naylor, Bradley C and Porter, Michael T and Wilson, Elise and Herring, Adam and Lofthouse, Spencer and Hannemann, Austin and Piccolo, Stephen R and Rockwood, Alan L and Price, John C},
-title = {DeuteRater: a tool for quantifying peptide isotope precision and kinetic proteomics},
-journal = {Bioinformatics},
-volume = {33},
-number = {10},
-pages = {1514-1520},
-year = {2017},
-doi = {10.1093/bioinformatics/btx009},
-URL = {http://dx.doi.org/10.1093/bioinformatics/btx009},
-eprint = {/oup/backfile/content_public/journal/bioinformatics/33/10/10.1093_bioinformatics_btx009/3/btx009.pdf}
+    author = {Naylor, Bradley C and Porter, Michael T and Wilson, Elise and Herring, Adam and Lofthouse, Spencer and Hannemann, Austin and Piccolo, Stephen R and Rockwood, Alan L and Price, John C},
+    title = {DeuteRater: a tool for quantifying peptide isotope precision and kinetic proteomics},
+    journal = {Bioinformatics},
+    volume = {33},
+    number = {10},
+    pages = {1514-1520},
+    year = {2017},
+    doi = {10.1093/bioinformatics/btx009},
+    URL = {http://dx.doi.org/10.1093/bioinformatics/btx009},
+    eprint = {/oup/backfile/content_public/journal/bioinformatics/33/10/10.1093_bioinformatics_btx009/3/btx009.pdf}
 }
