+++ README.md
@@ -3,6 +3,7 @@
 [![Build Status](https://travis-ci.org/emptyport/js-emass.svg?branch=master)](https://travis-ci.org/emptyport/js-emass)
 [![codecov](https://codecov.io/gh/emptyport/js-emass/branch/master/graph/badge.svg)](https://codecov.io/gh/emptyport/js-emass)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![status](http://joss.theoj.org/papers/cbc6dcaf1110a965cc8c4f1efc6a0963/status.svg)](http://joss.theoj.org/papers/cbc6dcaf1110a965cc8c4f1efc6a0963)
 
 
 
@@ -22,11 +23,14 @@ This implementation is functionally identical to the original emass, but it will
 View this module on [npm](https://www.npmjs.com/package/js-emass).
 
 ## Installation
-npm install js-emass --save
+`npm install js-emass --save`
+
+Additionally you can run `npm install molecular-formula --save` to be able to directly convert a formula such as "H2O" to the correct format for js-emass.
 
 ## Usage
 
 ### Quickstart
+You can run the following commands interactively in the terminal by running `node` and then typing the commands, or you can place everything in a file called `script.js` for example and run it with `node script.js`. This requires that [Node.js](https://nodejs.org) be installed on your system.
 ```javascript
 var emass_lib = require('js-emass');
 var molFormula = require('molecular-formula');
@@ -45,6 +49,8 @@ for(var i=0; i<isotopomers.length; i++) {
 ```
 Output:
 ```
+{ C: 100 }
+
 Mass: 1200, Abundance: 0.92457944
 Mass: 1201.003355, Abundance: 1
 Mass: 1202.00671, Abundance: 0.53537855
@@ -54,6 +60,60 @@ Mass: 1205.016774, Abundance: 0.01030258
 Mass: 1206.020129, Abundance: 0.00176431
 Mass: 1207.023484, Abundance: 0.00025625
 ```
+
+### Stable Isotope Labelling Example
+```javascript
+var emass_lib = require("js-emass");
+var molFormula = require('molecular-formula');
+
+var emassNatural = new emass_lib();
+var emassLabelled = new emass_lib();
+
+// Here we are adjusting our naturally occurring ratios.
+// In a rough sense we are taking one out of every 10 12C atoms and replacing it with a 13C atom.
+emassLabelled.addCustomIsotopes('C', [
+  {
+    "Mass": 12,
+    "Abundance": 0.8893
+  },
+  {
+    "Mass": 13.00335483507,
+    "Abundance": 0.1107
+  }
+]);
+
+var formula = new molFormula('C10');
+
+var isotopesNatural = emassNatural.calculate(formula.composition, 0);
+console.log("Natural isotopic envelope");
+console.log(isotopesNatural);
+
+console.log("\n");
+console.log("Labelled isotopic envelope");
+var isotopesLabelled = emassLabelled.calculate(formula.composition, 0);
+console.log(isotopesLabelled);
+```
+
+Output
+```
+Natural isotopic envelope
+[ { Mass: 120, Abundance: 1 },
+  { Mass: 121.003355, Abundance: 0.10815728 },
+  { Mass: 122.00671, Abundance: 0.0052641 },
+  { Mass: 123.010065, Abundance: 0.00015183 } ]
+
+
+Labelled isotopic envelope
+[ { Mass: 120, Abundance: 0.80334237 },
+  { Mass: 121.003355, Abundance: 1 },
+  { Mass: 122.00671, Abundance: 0.56015968 },
[…]
