### 2018-08-07 06cc662
fixing grammar in paper.md
