+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 # Summary
 js-emass is a JavaScript module that quickly and accurately calculates the isotopic envelope of a molecule. Isotopic envelopes are commonly encountered in the field of mass spectrometry which identifies molecules by their mass fingerprint. A mass spectrum is generated by measuring many copies of a molecule. The natural variations of isotopes then leads to a splitting of the signal. Instead of measuring a single mass, the instrument measures multiple masses which are then dubbed the isotopic envelope.
 
-The intended audience of js-emass are mass spectrometrists, particularly those using stable isotope labelling. In a typical stable isotope labelling experiment, the natural abundance of an isotope - such as deuterium, carbon 13, or nitrogen 15 - is enriched. By sampling at different timepoints, the rate of integration of the heavy isotope can be determined by constructing a series of isotopic envelopes between un-labelled and fully labelled to determine where on the curve each molecule lies. A molecule that experiences rapid turnover will be fully labelled with heavy isotopes quicker than a slowly synthesized molecule.
+The intended audience of js-emass are mass spectrometrists, particularly those using stable isotope labelling. In a typical stable isotope labelling experiment, the natural abundance of an isotope - such as deuterium, carbon 13, or nitrogen 15 - is enriched. By sampling at different timepoints, the rate of integration of the heavy isotope can be determined by constructing a series of isotopic envelopes between the un-labelled and fully labelled states to determine where on the curve each molecule lies. A molecule that experiences rapid turnover will be fully labelled with heavy isotopes quicker than a slowly synthesized molecule.
 
 The original emass algorithm [@ROCKWOOD2006415] was implemented in C++ as an interactive command line program. Dependencies on Windows libraries and the interactive nature of the program made it difficult to reuse. This implementation aims to be easy to use and fully configurable.
 
