### 2018-08-07 #869 @emptyport
I just fixed a small grammatical issue with the paper and now I think everything is good.
