### 2018-08-07 #869 @bovee
@emptyport Thanks for those changes! And sorry, but can you explicitly link `js-emass` to the problem in the second paragraph? Even something like:
  > ... can be determined by constructing a series of isotopic envelopes _with `js-emass`_ ...

would go a long way towards establishing the need for this package specifically (in addition to the text you already have about why existing solutions aren't as easy-to-use).

Other than that, I think this is great and I'm ready to sign off.
