+++ paper/paper.md
@@ -19,10 +19,10 @@ bibliography: paper.bib
 # Summary
 js-emass is a JavaScript module that quickly and accurately calculates the isotopic envelope of a molecule. Isotopic envelopes are commonly encountered in the field of mass spectrometry which identifies molecules by their mass fingerprint. A mass spectrum is generated by measuring many copies of a molecule. The natural variations of isotopes then leads to a splitting of the signal. Instead of measuring a single mass, the instrument measures multiple masses which are then dubbed the isotopic envelope.
 
-The intended audience of js-emass are mass spectrometrists, particularly those using stable isotope labelling. In a typical stable isotope labelling experiment, the natural abundance of an isotope - such as deuterium, carbon 13, or nitrogen 15 - is enriched. By sampling at different timepoints, the rate of integration of the heavy isotope can be determined by constructing a series of isotopic envelopes between the un-labelled and fully labelled states to determine where on the curve each molecule lies. A molecule that experiences rapid turnover will be fully labelled with heavy isotopes quicker than a slowly synthesized molecule.
+The intended audience of js-emass are mass spectrometrists, particularly those using stable isotope labelling. In a typical stable isotope labelling experiment, the natural abundance of an isotope - such as deuterium, carbon 13, or nitrogen 15 - is enriched. By sampling at different timepoints, the rate of integration of the heavy isotope can be determined by using js-emass to construct a series of isotopic envelopes between the un-labelled and fully labelled states to determine where on the curve each molecule lies. A molecule that experiences rapid turnover will be fully labelled with heavy isotopes quicker than a slowly synthesized molecule.
 
 The original emass algorithm [@ROCKWOOD2006415] was implemented in C++ as an interactive command line program. Dependencies on Windows libraries and the interactive nature of the program made it difficult to reuse. This implementation aims to be easy to use and fully configurable.
 
-All isotope data is fully up to date from NIST, but the user is still provided the option to provide custom isotope data for cases in which the natural abundances have been altered by the addition of heavy isotopes [@DEUTERATER]. This module also allows gives the user control over the charge state and various processing parameters.
+All isotope data is fully up to date from NIST, but the user is still provided the option to provide custom isotope data for cases in which the natural abundances have been altered by the addition of heavy isotopes [@DEUTERATER]. This module also allows the user control over the charge state and various processing parameters.
 
 # References
