### 2018-08-22 #869 @emptyport
@arfon all done. [Here is a link to the software in Zenodo](https://zenodo.org/badge/latestdoi/131650407
) and the DOI is 10.5281/zenodo.1402193
